@@ -15,7 +15,8 @@ import { Container } from '@mui/system';
 import Box from '@mui/material/Box';
 import Typography from '@mui/material/Typography';
 import Modal from 'react-bootstrap/Modal';
-import { Divider } from '@material-ui/core';
+import Divider from '@mui/material/Divider';
+
 
 function not(a, b) {
   return a.filter((value) => b.indexOf(value) === -1);
@@ -226,6 +227,6 @@ export default function ListaDeGrupos({
 ListaDeGrupos.propTypes = {
   idCurso: PropTypes.string.isRequired,
   show: PropTypes.bool.isRequired,
-  closeModal: PropTypes.element.isRequired,
-  setGruposParaTrabajo: PropTypes.element.isRequired, // Ajusta el tipo y la obligatoriedad según tu lógica de uso
+  closeModal: PropTypes.func.isRequired,// Cambia el tipo a func
+  setGruposParaTrabajo: PropTypes.func.isRequired, // Ajusta el tipo y la obligatoriedad según tu lógica de uso
 };
