@@ -1,11 +1,17 @@
-import { createMuiTheme } from '@material-ui/core';
+import { createTheme } from '@mui/material/styles';
 
 // Acá podrían reemplazarse los colores y tipografías del tema, y eso se va a reflejar en toda la aplicación.
-const theme = createMuiTheme({
-  palette: {
-    primary: {
-      // Como ejemplo, cambiamos el primary por "verde UNaHur".
-      main: '#009673',
+const theme = createTheme({
+  components: {
+    MuiButton: {
+      styleOverrides: {
+        root: {
+          backgroundColor: '#c5e1a5',
+          '&:hover': {
+            backgroundColor: '#b0d38a',
+          },
+        },
+      },
     },
   },
 });
